+++ test/debug_agent_test.cpp
@@ -35,11 +35,15 @@
 extern void VectorAddNormalTest ();
 extern void VectorAddDebugTrapTest ();
 extern void VectorAddMemoryFaultTest ();
+extern void SnapshotCodeObjOnLoadTest ();
+extern void SaveCodeObjectTest ();
 
 static void PrintTestInfo (const char *header);
 static void RunVectorAddDebugTrapTest ();
 static void RunVectorAddNormalTest ();
 static void RunVectorAddMemoryFaultTest ();
+static void RunSnapshotCodeObjOnLoadTest ();
+static void RunSaveCodeObjectTest ();
 
 int
 main (int argc, char *argv[])
@@ -51,6 +55,8 @@ main (int argc, char *argv[])
       run_test_list.push_back (0);
       run_test_list.push_back (1);
       run_test_list.push_back (2);
+      run_test_list.push_back (3);
+      run_test_list.push_back (4);
     }
   else
     {
@@ -75,6 +81,12 @@ main (int argc, char *argv[])
           break;
         case 2:
           RunVectorAddMemoryFaultTest ();
+          break;
+        case 3:
+          RunSnapshotCodeObjOnLoadTest ();
+          break;
+        case 4:
+          RunSaveCodeObjectTest ();
           break;
         default:
           std::cout << "  *** Invalid Test ID ***" << std::endl;
@@ -159,4 +171,34 @@ RunVectorAddMemoryFaultTest ()
     }
 
   PrintTestInfo ("VectorAddMemoryFaultTest end");
+}
+
+static void
+RunSnapshotCodeObjOnLoadTest ()
+{
+  PrintTestInfo ("VectorAddMemoryFaultTest start");
+
+  int deviceCount;
+  hipError_t err = hipGetDeviceCount (&deviceCount);
+  TEST_ASSERT (err == hipSuccess, "hipGetDeviceCount");
+
+  for (int i = 0; i < deviceCount; ++i)
+    {
+      err = hipSetDevice (i);
+      TEST_ASSERT (err == hipSuccess, "hipSetDevice");
+
+      SnapshotCodeObjOnLoadTest ();
+
+      err = hipDeviceReset ();
+      TEST_ASSERT (err == hipSuccess, "hipDeviceReset");
+    }
+
+  PrintTestInfo ("VectorAddMemoryFaultTest end");
+}
+
+static void
+RunSaveCodeObjectTest ()
+{
+  PrintTestInfo ("SaveCodeObjectTest start");
+  SaveCodeObjectTest ();
 }