+++ test/run-test.py
@@ -2,6 +2,8 @@ import os
 import re
 import sys
 import inspect
+import tempfile
+import unittest.mock
 from subprocess import Popen, PIPE
 
 
@@ -27,7 +29,6 @@ else:
     else:
         os.environ["LD_LIBRARY_PATH"] += ":" + agent_library_directory
     os.environ["HSA_TOOLS_LIB"] = "librocm-debug-agent.so.2"
-    os.environ["ROCM_DEBUG_AGENT_OPTIONS"] = "-p"
     os.chdir(test_binary_directory)
     # pre test to check if librocm-debug-agent.so.2 can be found
     p = Popen(['./rocm-debug-agent-test', '0'], stdout=PIPE, stderr=PIPE)
@@ -60,8 +61,8 @@ def check_test_1():
     print("Starting rocm-debug-agent test 1")
 
     #TODO: use regular expressions instead of strings
-    check_list = ['HSA_STATUS_ERROR_EXCEPTION: An HSAIL operation resulted in a hardware exception\.',
-                  '\(stopped, reason: ASSERT_TRAP\)',
+    check_list = ['HSA_STATUS_ERROR_EXCEPTION: An HSAIL operation resulted in a hardware exception\\.',
+                  '\\(stopped, reason: ASSERT_TRAP\\)',
                    'exec: (00000000)?00000001',
 #                  'status: 00012061',
 #                  'trapsts: 20000000',
@@ -69,7 +70,7 @@ def check_test_1():
                   's0:',
                   'v0:',
                   '0x0000: 22222222 11111111', # First uint64_t in LDS is '1111111122222222'
-                  'Disassembly for function vector_add_assert_trap\(int\*, int\*, int\*\)'
+                  'Disassembly for function vector_add_assert_trap\\(int\\*, int\\*, int\\*\\)'
 #                  'vector_add_assert_trap.cpp:', # Debug info may not be available on some older distributions
 #                  '53          __builtin_trap ();', # Source files not always available (When install tests from package)
 #                  's_trap 2'
@@ -105,7 +106,7 @@ def check_test_2():
     check_list = [
 #                  'System event \(HSA_AMD_GPU_MEMORY_FAULT_EVENT\)',
 #                  'Faulting page: 0x',
-                  '\(stopped, reason: MEMORY_VIOLATION\)',
+                  '\\(stopped, reason: MEMORY_VIOLATION\\)',
                   'exec: (ffffffff)?ffffffff',
 #                  'status: 00012461',
 #                  'trapsts: 30000100',
@@ -113,7 +114,7 @@ def check_test_2():
                   's0:',
                   'v0:',
                   '0x0000: 22222222 11111111', # First uint64_t in LDS is '1111111122222222'
-                  'Disassembly for function vector_add_memory_fault\(int\*, int\*, int\*\)'
+                  'Disassembly for function vector_add_memory_fault\\(int\\*, int\\*, int\\*\\)'
 #                  'vector_add_memory_fault.cpp:', Debug info may not be available on some older distributions
 #                  'global_store_dword' # Without precise memory, we can't guarantee that
                   ]
@@ -138,10 +139,95 @@ def check_test_2():
 
     return all_output_string_found
 
+# test 3: snapshot code object on load
+def check_test_3():
+    print("Starting rocm-debug-agent test 3")
+
+    p = Popen(['./rocm-debug-agent-test', '3'], stdout=PIPE, stderr=PIPE)
+    output, err = p.communicate()
+    out_str = output.decode('utf-8')
+    err_str = err.decode('utf-8')
+
+    found_error = False
+
+    # If the debug agent did not capture the code object on load, it should
+    # not be able to open it on exception, leading to the following warning:
+    #
+    #    rocm-debug-agent: warning: elf_getphdrnum failed for `memory://226967#offset=0x1651d4f0&size=3456'
+    #    rocm-debug-agent: warning: could not open code_object_1
+    if "could not open code_object" in  err_str:
+        found_error = True
+
+    # If the code object was not properly loaded, we should not have any
+    # disassembly in the output
+    if ("Disassembly:\n" not in err_str
+            and "Disassembly for function kernel_abort:\n" not in err_str):
+        found_error = True
+
+    if (found_error):
+        print("rocm-debug-agent test print out.")
+        print(out_str)
+        print("rocm-debug-agent test error message.")
+        print(err_str)
+
+    return not found_error
+
+# test 3: save code object on disk
+def check_test_4():
+    print("Starting rocm-debug-agent test 4")
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        with unittest.mock.patch.dict(os.environ, {"ROCM_DEBUG_AGENT_OPTIONS":
+                                                   f"-p --save-code-objects={tmpdir}"}):
+
+            p = Popen(['./rocm-debug-agent-test', '4'], stdout=PIPE, stderr=PIPE)
+            p.wait()
+
+            code_objects = os.listdir(tmpdir)
+            if (len(code_objects) == 0):
+                print(f"No code object found in {tmpdir}")
+                return False
+
+            # There should be 2 code objects which have the same address in
+            # memory and the same size (but might have different load
+            # addressed).  If the name did not have a "N_" prefix (N being a
+            # unique ID), those would be saved with the same file name, so
+            # the second saved code object would override the first one.
+            #
+            # This means that we should see 2 memory code objects:
+            # - 1_memory___PID_offset_OFF_size_SIZE
+            # - 2_memory___PID_offset_OFF_size_SIZE
+            #
+            # Trimming the "N_memory" prefix should give the same value for
+            # both, so a set containing those suffixes should have less
+            # elements than the set of initial names.
+            mem_cos = [co for co in code_objects if "memory___" in co]
+            if len(mem_cos) == len({co.split("memory")[1] for co in mem_cos}):
+                print("Unexpected number of unique code objects")
+                print("List of code objects:\n\t{}"
+                      "".format("\n\t".join(code_objects)))
+                return False
+
+            return True
+
 test_success = True
-test_success &= check_test_0()
-test_success &= check_test_1()
-test_success &= check_test_2()
+
+for deferred_loading in (None, "1", "0"):
+    with unittest.mock.patch.dict ('os.environ'):
+        if deferred_loading is None:
+            print(f"### Testing without HIP_ENABLE_DEFERRED_LOADING")
+            if "HIP_ENABLE_DEFERRED_LOADING" in os.environ:
+                del os.environ["HIP_ENABLE_DEFERRED_LOADING"]
+        else:
+            print(f"### Testing with HIP_ENABLE_DEFERRED_LOADING={deferred_loading}")
+            os.environ["HIP_ENABLE_DEFERRED_LOADING"] = deferred_loading
+
+        test_success &= check_test_0()
+        test_success &= check_test_1()
+        test_success &= check_test_2()
+        test_success &= check_test_3()
+        test_success &= check_test_4()
+
 if (test_success):
     print("rocm-debug-agent test Pass!")
 else: