+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Full documentation for the ROCR Debug Agent library is available at
 [rocm.docs.amd.com/rocr_debug_agent](https://rocm.docs.amd.com/projects/rocr_debug_agent/en/latest/).
 
+## ROCR Debug Agent 2.1.0 for ROCm 7.0
+
+### Added
+- Added the `-e` and `--precise-alu-exceptions` flags to enable precise
+  ALU exceptions reporting on supported configurations.
+
 ## ROCR Debug Agent 2.0.4 for ROCm 6.4
 
 ### Added